Ticket Statuses from DB
First swing at this to share my progress, isn't ready to merge yet but would appreciate thoughts

+++ post/ticket.php
@@ -11,9 +11,9 @@ if (isset($_POST['add_ticket'])) {
     $client_id = intval($_POST['client']);
     $assigned_to = intval($_POST['assigned_to']);
     if ($assigned_to == 0) {
-        $ticket_status = 'New';
+        $ticket_status = $config_ticket_status_id_new;
     } else {
-        $ticket_status = 'Open';
+        $ticket_status = $config_ticket_status_id_open;
     }
     $contact = intval($_POST['contact']);
     $subject = sanitizeInput($_POST['subject']);
@@ -82,6 +82,7 @@ if (isset($_POST['add_ticket'])) {
         $ticket_details = mysqli_escape_string($mysqli, $row['ticket_details']);
         $ticket_priority = sanitizeInput($row['ticket_priority']);
         $ticket_status = sanitizeInput($row['ticket_status']);
+        $ticket_status_name = sanitizeInput(getTicketStatusName($row['ticket_status']));
         $client_id = intval($row['ticket_client_id']);
         $ticket_created_by = intval($row['ticket_created_by']);
         $ticket_assigned_to = intval($row['ticket_assigned_to']);
@@ -330,8 +331,15 @@ if (isset($_POST['assign_ticket'])) {
     $ticket_id = intval($_POST['ticket_id']);
     $assigned_to = intval($_POST['assigned_to']);
     $ticket_status = sanitizeInput($_POST['ticket_status']);
+
+    // TODO: Remove this legacy if
     if ($ticket_status == 'New' && $assigned_to !== 0) {
-        $ticket_status = 'Open';
+        $ticket_status = $config_ticket_status_id_open;
+    }
+
+    // New
+    if ($ticket_status == $config_ticket_status_id_new && $assigned_to !== 0) {
+        $ticket_status = $config_ticket_status_id_open;
     }
 
     // Allow for un-assigning tickets
@@ -356,7 +364,7 @@ if (isset($_POST['assign_ticket'])) {
     }
 
     // Get & verify ticket details
-    $ticket_details_sql = mysqli_query($mysqli, "SELECT ticket_prefix, ticket_number, ticket_subject, ticket_client_id, client_name FROM tickets LEFT JOIN clients ON ticket_client_id = client_id WHERE ticket_id = '$ticket_id' AND ticket_status != 'Closed'");
+    $ticket_details_sql = mysqli_query($mysqli, "SELECT ticket_prefix, ticket_number, ticket_subject, ticket_client_id, client_name FROM tickets LEFT JOIN clients ON ticket_client_id = client_id WHERE ticket_id = '$ticket_id' AND ticket_status != $config_ticket_status_id_closed AND ticket_status != 'Closed'");
     $ticket_details = mysqli_fetch_array($ticket_details_sql);
 
     $ticket_prefix = sanitizeInput($ticket_details['ticket_prefix']);
@@ -434,7 +442,7 @@ if (isset($_GET['delete_ticket'])) {
     $ticket_status = sanitizeInput($row['ticket_status']);
     $client_id = intval($row['ticket_client_id']);
 
-    if ($ticket_status !== 'Closed') {
+    if ($ticket_status !== $config_ticket_status_id_closed && $ticket_status !== 'Closed') {
         mysqli_query($mysqli, "DELETE FROM tickets WHERE ticket_id = $ticket_id");
 
         // Delete all ticket replies
@@ -479,7 +487,7 @@ if (isset($_POST['bulk_assign_ticket'])) {
             $client_id = intval($row['ticket_client_id']);
 
             if ($ticket_status == 'New' && $assigned_to !== 0) {
-                $ticket_status = 'Open';
+                $ticket_status = $config_ticket_status_id_open;
             }
 
             // Allow for un-assigning tickets
@@ -621,7 +629,7 @@ if (isset($_POST['bulk_close_tickets'])) {
 
             $ticket_prefix = sanitizeInput($row['ticket_prefix']);
             $ticket_number = intval($row['ticket_number']);
-            $ticket_status = sanitizeInput($row['ticket_status']);
+            $ticket_status = sanitizeInput(getTicketStatusName($row['ticket_status']));
             $ticket_subject = sanitizeInput($row['ticket_subject']);
             $current_ticket_priority = sanitizeInput($row['ticket_priority']);
             $client_id = intval($row['ticket_client_id']);
@@ -713,6 +721,7 @@ if (isset($_POST['bulk_ticket_reply'])) {
     // POST variables
     $ticket_reply = mysqli_escape_string($mysqli, $_POST['bulk_reply_details']);
     $ticket_status = sanitizeInput($_POST['bulk_status']);
+    $ticket_status_name = sanitizeInput(getTicketStatusName($row['ticket_status']));
     $private_note = intval($_POST['bulk_private_reply']);
     if ($private_note == 1) {
         $ticket_reply_type = 'Internal';
@@ -780,7 +789,7 @@ if (isset($_POST['bulk_ticket_reply'])) {
                 if (filter_var($contact_email, FILTER_VALIDATE_EMAIL)) {
 
                     $subject = "Ticket update - [$ticket_prefix$ticket_number] - $ticket_subject";
-                    $body = "<i style=\'color: #808080\'>##- Please type your reply above this line -##</i><br><br>Hello $contact_name,<br><br>Your ticket regarding $ticket_subject has been updated.<br><br>--------------------------------<br>$ticket_reply<br>--------------------------------<br><br>Ticket: $ticket_prefix$ticket_number<br>Subject: $ticket_subject<br>Status: $ticket_status<br>Portal: https://$base_url/portal/ticket.php?id=$ticket_id<br><br>--<br>$company_name - Support<br>$from_email<br>$company_phone";
+                    $body = "<i style=\'color: #808080\'>##- Please type your reply above this line -##</i><br><br>Hello $contact_name,<br><br>Your ticket regarding $ticket_subject has been updated.<br><br>--------------------------------<br>$ticket_reply<br>--------------------------------<br><br>Ticket: $ticket_prefix$ticket_number<br>Subject: $ticket_subject<br>Status: $ticket_status_name<br>Portal: https://$base_url/portal/ticket.php?id=$ticket_id<br><br>--<br>$company_name - Support<br>$from_email<br>$company_phone";
 
                     $data = [];
 
@@ -842,6 +851,7 @@ if (isset($_POST['add_ticket_reply'])) {
     $ticket_id = intval($_POST['ticket_id']);
     $ticket_reply = mysqli_real_escape_string($mysqli, $_POST['ticket_reply']);
     $ticket_status = sanitizeInput($_POST['status']);
+    $ticket_status_name = sanitizeInput(getTicketStatusName($row['ticket_status']));
     // Handle the time inputs for hours, minutes, and seconds
     $hours = intval($_POST['hours']);
     $minutes = intval($_POST['minutes']);
@@ -869,7 +879,7 @@ if (isset($_POST['add_ticket_reply'])) {
     // Update Ticket Last Response Field
     mysqli_query($mysqli, "UPDATE tickets SET ticket_status = '$ticket_status' WHERE ticket_id = $ticket_id");
 
-    if ($ticket_status == 'Closed') {
+    if ($ticket_status == $config_ticket_status_id_closed || $ticket_status == 'Closed') {
         mysqli_query($mysqli, "UPDATE tickets SET ticket_closed_at = NOW() WHERE ticket_id = $ticket_id");
     }
 
@@ -909,15 +919,15 @@ if (isset($_POST['add_ticket_reply'])) {
 
             // Slightly different email subject/text depending on if this update closed the ticket or not
 
-            if ($ticket_status == 'Closed') {
+            if ($ticket_status == $config_ticket_status_id_closed || $ticket_status == 'Closed') {
                 $subject = "Ticket closed - [$ticket_prefix$ticket_number] - $ticket_subject | (do not reply)";
                 $body = "Hello $contact_name,<br><br>Your ticket regarding $ticket_subject has been closed.<br><br>--------------------------------<br>$ticket_reply<br>--------------------------------<br><br>We hope the request/issue was resolved to your satisfaction. If you need further assistance, please raise a new ticket using the below details. Please do not reply to this email. <br><br>Ticket: $ticket_prefix$ticket_number<br>Subject: $ticket_subject<br>Portal: https://$config_base_url/portal/ticket.php?id=$ticket_id<br><br>--<br>$company_name - Support<br>$config_ticket_from_email<br>$company_phone";
-            } elseif ($ticket_status == 'Auto Close') {
+            } elseif ($ticket_status == $config_ticket_status_id_autoclose || $ticket_status == 'Auto Close') {
                 $subject = "Ticket update - [$ticket_prefix$ticket_number] - $ticket_subject | (pending closure)";
-                $body = "<i style=\'color: #808080\'>##- Please type your reply above this line -##</i><br><br>Hello $contact_name,<br><br>Your ticket regarding $ticket_subject has been updated and is pending closure.<br><br>--------------------------------<br>$ticket_reply<br>--------------------------------<br><br>If your request/issue is resolved, you can simply ignore this email. If you need further assistance, please respond to let us know!  <br><br>Ticket: $ticket_prefix$ticket_number<br>Subject: $ticket_subject<br>Status: $ticket_status<br>Portal: https://$config_base_url/portal/ticket.php?id=$ticket_id<br><br>--<br>$company_name - Support<br>$config_ticket_from_email<br>$company_phone";
+                $body = "<i style=\'color: #808080\'>##- Please type your reply above this line -##</i><br><br>Hello $contact_name,<br><br>Your ticket regarding $ticket_subject has been updated and is pending closure.<br><br>--------------------------------<br>$ticket_reply<br>--------------------------------<br><br>If your request/issue is resolved, you can simply ignore this email. If you need further assistance, please respond to let us know!  <br><br>Ticket: $ticket_prefix$ticket_number<br>Subject: $ticket_subject<br>Status: $ticket_status_name<br>Portal: https://$config_base_url/portal/ticket.php?id=$ticket_id<br><br>--<br>$company_name - Support<br>$config_ticket_from_email<br>$company_phone";
             } else {
                 $subject = "Ticket update - [$ticket_prefix$ticket_number] - $ticket_subject";
-                $body = "<i style=\'color: #808080\'>##- Please type your reply above this line -##</i><br><br>Hello $contact_name,<br><br>Your ticket regarding $ticket_subject has been updated.<br><br>--------------------------------<br>$ticket_reply<br>--------------------------------<br><br>Ticket: $ticket_prefix$ticket_number<br>Subject: $ticket_subject<br>Status: $ticket_status<br>Portal: https://$config_base_url/portal/ticket.php?id=$ticket_id<br><br>--<br>$company_name - Support<br>$config_ticket_from_email<br>$company_phone";
+                $body = "<i style=\'color: #808080\'>##- Please type your reply above this line -##</i><br><br>Hello $contact_name,<br><br>Your ticket regarding $ticket_subject has been updated.<br><br>--------------------------------<br>$ticket_reply<br>--------------------------------<br><br>Ticket: $ticket_prefix$ticket_number<br>Subject: $ticket_subject<br>Status: $ticket_status_name<br>Portal: https://$config_base_url/portal/ticket.php?id=$ticket_id<br><br>--<br>$company_name - Support<br>$config_ticket_from_email<br>$company_phone";
             }
 
             $data = [];
